cleanup httpclient dir
