+++ docs/release_notes/v1.x/v1.0.10.md
@@ -165,7 +165,7 @@ v1.0.10 is a major release that hardens the platform's security posture, simplif
 * **Storage layer refactor** — Major refactor of the storage subsystem:
   * `TokenStoreService.KVScanPrefix` now decrypts values (previously returned encrypted ciphertext).
   * Removed dead `TextScrubber` dependency from `ExecutionVaultService`.
-  * Chaos test infrastructure moved to `internal/test/chaos/`.
+  * Chaos test infrastructure moved to `test/chaos/`.
 
 * **Gateway architecture** — Decomposed gateway HTTP handling into dedicated controllers:
   * `AuthController` — authentication, enrollment, and session management.
