+++ internal/cli/cmd/chaos.go
@@ -19,7 +19,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/g8e-ai/g8e/internal/constants"
-	"github.com/g8e-ai/g8e/internal/test/chaos"
+	"github.com/g8e-ai/g8e/test/chaos"
 )
 
 var (
