+++ internal/cli/cmd/chaos_test.go
@@ -16,7 +16,7 @@ package cmd
 import (
 	"testing"
 
-	"github.com/g8e-ai/g8e/internal/test/chaos"
+	"github.com/g8e-ai/g8e/test/chaos"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
