+++ internal/httpclient/httpclient.go
@@ -15,6 +15,7 @@ package httpclient
 
 import (
 	"crypto/tls"
+	"encoding/json"
 	"fmt"
 	"net"
 	"net/http"
@@ -207,4 +208,36 @@ func WebSocketDialerWithTLSConfigAndServerName(tlsConfig *certs.TLSConfig, serve
 		TLSClientConfig:  tlsCfg,
 		HandshakeTimeout: DefaultTLSTimeout,
 	}, nil
+}
+
+// ExtractErrorMessage returns a human-readable error string from a raw JSON
+// `error` field produced by client, accepting either:
+//   - a plain JSON string: "some error"
+//   - the standard client error envelope object: {"code": "...", "message": "...", ...}
+//
+// g8eo HTTP response structs should model `error` as json.RawMessage rather
+// than `string`, and call this helper when surfacing the error to the user.
+// Modeling it as a bare `string` causes a silent decode failure whenever the
+// server returns the object form, masking the real server error.
+func ExtractErrorMessage(raw json.RawMessage) string {
+	if len(raw) == 0 {
+		return ""
+	}
+	var s string
+	if err := json.Unmarshal(raw, &s); err == nil {
+		return s
+	}
+	var obj struct {
+		Message string `json:"message"`
+		Code    string `json:"code"`
+	}
+	if err := json.Unmarshal(raw, &obj); err == nil {
+		if obj.Message != "" && obj.Code != "" {
+			return fmt.Sprintf("%s: %s", obj.Code, obj.Message)
+		}
+		if obj.Message != "" {
+			return obj.Message
+		}
+	}
+	return string(raw)
 }
